Revert "Merge tag 'MP_3.45.0_BCM21664Android_SystemRel_6.3.10' into sdb-linux-3.10.0-hawaii-java"
This reverts commit e5bd35b8dc7dce504e3490576a7fef26b687262d, reversing
changes made to 45e7c9a926d290562c7a3d77de2a7cd89561175d.

Reviewed-on: http://mps-gerrit.broadcom.com/45684
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>

+++ drivers/char/broadcom/fuse_log/output.c
@@ -293,10 +293,11 @@ static void WriteToLogDev_SDCARD(void)
 						   nFifo,
 						   &g_devWrParms.
 						   file->f_pos);
+
 				if (nWrite > 0)
 					BCMLOG_FifoRemove(&g_fifo, nWrite);
 
-				if ((nWrite < (int)nFifo)
+				if ((nWrite < nFifo)
 				    || ((Get_SDCARD_Available()) <
 					MTT_SD_RESERVED)) {
 					nFifo = 0;
